Turborepo, pnpm and tsup :fire:
This PR configures eas-monorepo-example with a turborepo build process. It also updates packages/ui to use tsup instead of builder-bob, sorry bob. And it configures the app to use pnpm. 

We still need to shamefully-hoist - (Should I feel shame? Because I do.)

But, overall it’s a really nice build pipeline.
```
$ pnpm install --shamefully-hoist
$ pnpm build # to build the workspace with turborepo
$ pnpm managed start
$ pnpm clean # to reset everything
```

Feels good, man. 

And sets it up the monorepo nicely for dropping in apps that need to run in parallel like an Express API, Next or Remix.